Refactor Document's file & folder structure
Rename src/common/Document to src/common/Page
