+++ src/components/common/Page/index.js
@@ -1,0 +1,4 @@
+export { default as Body } from './Body';
+export { default as Document } from './Document';
+export { default as Head } from './Head';
+export { default as Title } from './Title';
